Converted page to a class
